move downloud package to datatx/utils
