+++ pkg/rhttp/datatx/utils/download/range.go
